+++ packages/youtube_player_iframe/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added **Migrating to v3** section in the docs.
 - Listed `Swipe up/down to enter/exit fullscreen mode` as a feature in the docs.
 - Deprecated `autoPlay` param in **YoutubePlayerParams**, as it's no longer used.
+- Removed `hasPlayed`, `position` & `buffered` from **YoutubePlayerValue**, as these values can be accessed through **YoutubePlayerController**.
 
 # 3.0.2
 - Improved pub score.